Rename columns in CSV files to use "_id" suffix
There are many places in the CSV files where rows in one file are referred to in another, e.g. commodities defined in commodities.csv are referred to via a "commodity_name" column elsewhere. Agents and regions are the exception: they're referred to with "agent" and "region", respectively.

As "name" is a little ambiguous, I've used an "id" suffix instead. In files that provide a list of these IDs (e.g. agents.csv) the column is just "id" now, which is terser and more in keeping with what databases do.

Closes #101.