CI: Check that API documentation builds without warnings
I just noticed that `cargo doc` is emitting a warning for our code, but I didn't noticed until now, because, while we do build and the API documentation on pushes to `main`, at no point do we check for warnings.

This PR adds a workflow to do this for PRs. The docs workflow is also changed to fail on warnings; while this means that docs with warnings won't be published, it also means we'll get an email if we break the documentation on `main`.