Use consistent casing for enum-type fields in CSV files
As discussed in #120, we are not currently consistent in whether CSV column values are in lowercase (e.g. `limit_type`) or uppercase (e.g. `time_slice_level`). Let's be consistent throughout and just use lowercase everywhere.

While this won't affect the code as case will be normalised when reading in these values, we will want to have a policy for the sake of the example model(s) and documentation.

Closes #120.