Multiply `flow_cost` by -1 for inputs
From Adam:

The sign of the flow matters, but does not
influence whether the flow_cost is applied (i.e.
the cost can be applied to an input or an output).
But note that cost should be positive whenever a
positive cost is applied – and as input is
negative then we’ll have to multiply the flow_cost
by -1 to make sure impact on the objective
function is a positive cost. Make sense? We’ll
also have to document that this is how it works to
avoid confusions to the extent possible!