Fix examples/simple: Make last milestone year 2040
There is currently a problem in that the last milestone year is after all of the assets have been decommissioned, meaning that there will be no assets available for the last milestone year, which seems pretty nonsensical (albeit theoretically possible) and it causes the optimiser to barf. Let's fix this by changing the last milestone year.