Fix: Primary output should be updated if specified explicitly too
If the primary output for a set of process flows is specified explicitly (i.e. by putting `true` in the `is_primary_output` column), rather than inferred (which we can do if there is only one output flow) then the `is_primary_output` field is still left as `false` for that output flow, which is not correct. Fix this.