Don't load `settings.toml` in tests
If the `debug_model` setting is set to `true` in the user's `settings.toml` file, then the
regression tests will fail. The tests should not depend on user settings. Fix by changing a couple
of functions to accept an optional `Settings` struct, to allow tests to provide their own settings.

Closes #814.